refactor(Lotto): Lotto 패키지 변경
